requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
reportlab==4.0.7
Werkzeug==3.0.1
